requirements.txt
Django==1.11.10
pillow
mysqlclient

djangorestframework-jwt
djangorestframework
markdown
django-guardian
django-filter
coreapi

django-crispy-forms
django-reversion
django-formtools
future
httplib2
six
xlwt
xlsxwriter

requests
django-cors-headers
drf-extensions
django-redis
pycryptodome
raven
